Remove devices from `lora/devices`
The devices list is useful for manually adding, removing, and then
re-adding devices.
But if you are adding the devices automatically then it is less
important.

But most importantly, once the gateway device list reaches 2000 devices,
the gateway stops responding to further devices (without providing a
helpful response or warning).