Fix `devices::remove()`
EUIs can be formatted in different ways, so when checking for a device
in my list the EUI needs to be compared.

EUIs can be formatted in different ways, so when instructing the Gateway
to remove a device the EUI needs to be formatted exactly as the Gateway
formatted it.